Internal improvements (#98)
* Fix initial setup for running in internal github

* Run on PRs with code from PR

Also use internal DAX token

* Add tests subtask creation

* Add guard from creating multiple review tasks

* Add a few todos & fix subtask creation

* Reopen existing review subtasks if needed

* Ignore PRs which start with "Release: " prefix

This would fix https://app.asana.com/0/718967094418700/1204935661949375/f

* Add body of the Github PR in the Asana task

* Assign main PR task to the author of the PR

Subtasks are for reviewers

* Tweak PR text to include note on overwrite

* Try to remove end comments/links from description

* Add action output - task_id

* Add "result" to output

That can be either created or updated depending if a task was created or
updated :-)

* Switch to provide task_url parameter

* Add a few new username -> email map

* Try to close review subtask automatically

Rework things quite a bit and process more actions

* Refactor code to make things readable

* Add review_requested trigger action

* Ignore some users by default for Asana tasks

* Close remaining review tasks after PR is closed/merged

* Make Asana PR task subtask of referenced Asana

If people reference an Asana task let's automatically add our PR as a
subtask

* Update README instructions and description

* Add review_requested to example event triggers

This is needed to create subtasks properly

* Update examples and use concurrency to prevent duplicates

In https://app.asana.com/0/0/1205415895460206/1205415070380175/f this
task accidentally created duplicate Asana tasks because it ran for PR
and also for assignment.

This makes sure we serialize any asana sync runs which will prevent
duplicate tasks being created.

* Automatically close PR tasks on merge/close

With the caveat that the closing does not happen if the PR task has been
added to the REVIEW/RELEASE project because there we want to manually verify.

* Check access to project of mentioned Asana task

If we can't access it we don't add the task as a subtask of referenced
Asana task automatically but mention in the log how to enable that.

* Correct skipping for Release PRs

Original code incorrectly matched on title which wasn't the PR title but
the title of the Asana task.

* Add more logging to help with debugging

* (re)Open a PR review only if it is submitted and approved

Adding a code comment in a PR creates a "pull_request_review" event
without actual review.state payload. This meant that the original code
would actually open a review task in asana for random commenters.

* Stop assigning PR tasks to PR authors

Seems people are not too fond of this feature so let's disable assigning
to people. Keep the task, this way they get referenced in mentioned
Asana tasks.

* Fix duplicate asana task creation

When 2 workflows run back to back we hit Asana API inconsistency:
https://developers.asana.com/reference/searchtasksforworkspace#eventual-consistency

To fix this - if we don't find task URL with searchInWorkspace, we
follow up with findByProject and look through 100 tasks in the
destination project. New tasks *should* appear in the API response but
it's not garanteed.

* DELETE THIS

ignore searchInWorkspace to test things

* Remove concurrency setup

See: https://app.asana.com/0/1204969954708911/1205470363838442/f

See documentation/discussion in https://github.com/orgs/community/discussions/5435

* Add PRFields type

Will be used in later commits

* Add wait for PR to be created

We don't want to rely on concurrency to prevent double PR task creation.
Instead we only create PRs on "open" events and in all other cases
require that we find an existing PR task. Possibly spinning for a minute
or two waiting for it.

* Refactor the main code for easier reading

Hopefully the changes make sense. This is mostly moving existing code to
have less nesting and duplication of function calls.

* Tweak description of PRs a bit

* Close PRs based on state not just action

If another action happens after the PR is closed (i.e. someone edits a
description or checks a checkbox) the code would reopen the Asana task

* Add configuration mention in the README

* Fix searchInWorkspace calls

We used wrong variable to search so we always fell back to iterating
over tasks in project. This can fail for older tasks.

While at this start using PRFields custom type for simplification

* Don't bail if no task is found - create a new one

The difference is that we waited a sufficient time here so if we create
a new task we are very unlikely to create a duplicate.

* Fixup searchInWorkspace calls

Forgotten debugging call left has wreaked havoc...

* Tweak PR titles & review descriptions

See https://app.asana.com/0/1204969954708911/1205525315586001/f

* Add stacktrace to error messages

This makes it easier to debug if somethings goes wrong

* Print more debug info when creating review subtasks

This was caused by https://app.asana.com/0/908478224964033/1206310929197621/f

* Trim descriptions before submitting to Asana

Without the limit we can easily hit following when updating/creating the
task:
Error: Invalid Request
    at InvalidRequest.AsanaError (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/asana/lib/errors/error.js:4:1)
    at new InvalidRequest (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/asana/lib/errors/invalid_request.js:5:1)
    at Request._callback (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/asana/lib/dispatcher.js:263:1)
    at Request.self.callback (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/request/request.js:185:1)
    at Request.emit (node:events:527:28)
    at Request.<anonymous> (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/request/request.js:1154:1)
    at Request.emit (node:events:527:28)
    at IncomingMessage.<anonymous> (/mnt/ebs/opt/github-actions-runner1/_work/_actions/duckduckgo/action-asana-sync/master/webpack:/typescript-action/node_modules/request/request.js:1076:1)
    at Object.onceWrapper (node:events:641:28)
    at IncomingMessage.emit (node:events:539:35)
    at endReadableNT (node:internal/streams/readable:1345:12)
    at processTicksAndRejections (node:internal/process/task_queues:83:21)

* Add user map variable to workflow

* Fix action runners

* Add input descriptions

---------

Co-authored-by: Stanislav Ochotnický <stanislav@ochotnicky.com>
Co-authored-by: Stanislav Ochotnicky <sochotnicky@duckduckgo.com>
Co-authored-by: Dax Duck <dax@duckduckgo.com>